requirements.txt
# =============================================================================
# Aero-Watch — Mac-Compatible Dependencies
# =============================================================================
# Tested on: macOS 14+ (Sonoma), Apple Silicon (M1/M2/M3) and Intel

# --- PyTorch (Mac-native, includes MPS support) ---
torch>=2.1.0
torchvision>=0.16.0

# --- YOLOv8 ---
ultralytics>=8.1.0



# --- Computer Vision ---
opencv-python-headless>=4.8.0
Pillow>=10.0.0
scipy>=1.11.0

# --- API ---
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# --- Config & Utilities ---
pyyaml>=6.0
numpy>=1.24.0

# --- CLIP (zero-shot classification) ---
open-clip-torch>=2.20.0

# --- Multi-camera scaling ---
redis>=5.0.0
psycopg2-binary>=2.9.0

# --- Testing ---
pytest>=7.4.0
